move plots from composition to their own file
