+++ vivarium/experiments/glucose_phosphorylation.py
@@ -12,10 +12,10 @@ from vivarium.compartments.injected_glc_phosphorylation import (
     InjectedGlcPhosphorylation,
 )
 from vivarium.core.composition import (
-    plot_simulation_output,
     simulate_experiment,
     EXPERIMENT_OUT_DIR,
 )
+from vivarium.plots.output import plot_simulation_output
 from vivarium.core.emitter import (
     path_timeseries_from_data,
     timeseries_from_data,
