+++ vivarium/processes/injector.py
@@ -16,11 +16,11 @@ from __future__ import absolute_import, division, print_function
 import os
 
 from vivarium.core.composition import (
-    plot_simulation_output,
     simulate_process,
     TEST_OUT_DIR,
     save_timeseries,
 )
+from vivarium.plots.output import plot_simulation_output
 from vivarium.core.process import Process
 
 NAME = 'injector'
