+++ vivarium/processes/template_process.py
@@ -5,10 +5,9 @@ import os
 from vivarium.core.process import Process
 from vivarium.core.composition import (
     simulate_process_in_experiment,
-    plot_simulation_output,
     PROCESS_OUT_DIR,
 )
-
+from vivarium.plots.output import plot_simulation_output
 
 # a global NAME is used for the output directory and for the process name
 NAME = 'template'
